Feature/cleanup token invalid

+++ back-end/QLDT/Dtos/request/ClassReq.cs
@@ -52,7 +52,7 @@ namespace QLDT.Dtos.request
         public long MajorId { get; set; }
 
         [Required]
-        public int SoTinhChi { get; set; }
+        public double SoTinhChi { get; set; }
         public List<IFormFile> Attachments { get; set; } = new();
         public List<string>? OldFileIds { get; set; }
         public List<long> EmployeeIds { get; set; } = new();
